requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
requests>=2.31.0
python-pptx>=0.6.23
openpyxl>=3.1.2
kaleido>=0.2.1
